On branch main Changes to be committed: 	modified:   include/iptables/firewall.var 	modified:   lib/iptables/icore.ipv4 	modified:   lib/iptables/icore.ipv6 	new file:   lib/iptables/ocore.ipv4 	new file:   lib/iptables/ocore.ipv6 	modified:   sbin/start-firewall
Notes:
So far, I've simply past the test phase of no error codes on deployment. It's
current usefulness is, I promise it's not going to permit much! Next up, will
be a specific entry into filter-oTCP4 to permit whatever is directed to
debian repo servers as resolved on the fly!